refactor: move response values validators to a dedicated validators file
